requirements.txt
pycountry==23.12.11
pillow==9.4.0
